bot/requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
emoji==1.7.0
environs==9.5.0
frozenlist==1.3.0
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
hnswlib==0.6.2
idna==3.3
iniconfig==1.1.1
marshmallow==3.15.0
multidict==6.0.2
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.8.0
pyparsing==3.0.7
pytest==7.1.2
python-dotenv==0.20.0
python-rapidjson==1.6
pytz==2021.3
rsa==4.8
six==1.16.0
Telethon==1.24.0
toml==0.10.2
tomli==2.0.1
tritonclient==2.19.0
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
